+++ resources/fornecedor.py
@@ -1,5 +1,10 @@
 from flask_restful import Resource, reqparse
 from models.fornecedor_model import FornecedorModel
+from models.produto_model import ProdutoModel
+from resources.mensagem import fornecedorEmUso, erroSalvarFornecedor
+from resources.mensagem import fornecedorNaoEncontrado, fornecedorJaExiste
+from resources.mensagem import erroExcluirFornecedor, fornecedorExcluido
+from resources.mensagem import cnpjCpfJaExiste
 from flask_jwt_extended import jwt_required
 
 
@@ -25,7 +30,7 @@ class Fornecedor(Resource):
         fornecedor = FornecedorModel.find_produto(cod_fornecedor)
         if fornecedor:
             return fornecedor.json()
-        return {'mensagem': 'Produto não encontrado.'}, 404
+        return fornecedorNaoEncontrado
 
     @jwt_required
     def put(self, cod_fornecedor):
@@ -38,26 +43,28 @@ class Fornecedor(Resource):
             return fornecedor_encontrado.json(), 200
 
         if FornecedorModel.find_fornecedor(dados['cod_fornecedor']):
-            return {'mensagem': 'Fornecedor com código "{}" já existe.'
-                    .format(dados['cod_fornecedor'])}, 400
-
+            return fornecedorJaExiste(dados['cod_fornecedor'])
         fornecedor = FornecedorModel(cod_fornecedor, **dados)
         try:
             fornecedor.save_fornecedor()
         except ValueError:
-            return {'mensagem': 'Erro ao salvar o fornecedor.'}, 500
+            return erroSalvarFornecedor
         return fornecedor.json(), 201
 
     @jwt_required
     def delete(self, cod_fornecedor):
         fornecedor = FornecedorModel.find_fornecedor(cod_fornecedor)
+
+        if ProdutoModel.find_produto_fornecedor(cod_fornecedor):
+            return fornecedorEmUso(cod_fornecedor)
+
         if fornecedor:
             try:
                 fornecedor.delete_fornecedor()
             except ValueError:
-                return {'mensagem': 'Erro ao excluir o fornecedor.'}, 500
-            return {'mensagem': 'Fornecedor excluído.'}
-        return {'mensagem': 'Fornecedor não encontrado.'}, 404
+                return erroExcluirFornecedor
+            return fornecedorExcluido
+        return fornecedorNaoEncontrado
 
 
 class FornecedorCadastro(Resource):
@@ -66,13 +73,12 @@ class FornecedorCadastro(Resource):
         dados = Fornecedor.argumentos.parse_args()
 
         if FornecedorModel.find_fornecedor_cnpj_cpf(dados['cnpj_cpf']):
-            return {'mensagem': 'CNPJ/CPF "{}" já existe.'
-                    .format(dados['cnpj_cpf'])}, 400
+            return cnpjCpfJaExiste
 
         fornecedor = FornecedorModel(**dados)
 
         try:
             fornecedor.save_fornecedor()
         except ValueError:
-            return {'mensagem': 'Erro ao salvar o fornecedor.'}, 500
+            return erroSalvarFornecedor
         return fornecedor.json(), 200
